Fix -Wconversion warnings, and issues detected by those warnings

+++ Spatial_sorting/include/CGAL/Multiscale_sort.h
@@ -46,7 +46,7 @@ public:
 	typedef typename std::iterator_traits<RandomAccessIterator>::difference_type difference_type;
         RandomAccessIterator middle = begin;
         if (end - begin >= _threshold) {
-            middle = begin + difference_type ((end - begin) * _ratio);
+            middle = begin + difference_type (double(end - begin) * _ratio);
             this->operator() (begin, middle);
         }
         _sort (middle, end);
